requirements.txt
Django==5.2.4
djangorestframework==3.15.2
ezdxf==1.3.0
rarfile==4.1
coverage==7.9.2
django-cors-headers==4.3.1 